+++ 1.20.4/Common/src/main/java/fuzs/hangglider/client/handler/FovModifierHandler.java
@@ -1,13 +1,13 @@
 package fuzs.hangglider.client.handler;
 
-import fuzs.hangglider.helper.PlayerGlidingHelper;
+import fuzs.hangglider.init.ModRegistry;
 import fuzs.puzzleslib.api.event.v1.data.DefaultedFloat;
 import net.minecraft.world.entity.player.Player;
 
 public class FovModifierHandler {
 
     public static void onComputeFovModifier(Player player, DefaultedFloat fieldOfViewModifier) {
-        if (PlayerGlidingHelper.isGliding(player)) {
+        if (ModRegistry.GLIDING_CAPABILITY.get(player).isGliding()) {
             fieldOfViewModifier.mapFloat(f -> f * (player.isDescending() ? 1.1F : 1.05F));
         }
     }