+++ 1.20.4/Common/src/main/java/fuzs/hangglider/handler/GliderUseHandler.java
@@ -1,4 +1,0 @@
-package fuzs.hangglider.handler;
-
-public class GliderUseHandler {
-}