+++ 1.20.4/Common/src/main/java/fuzs/hangglider/init/ModRegistry.java
@@ -2,15 +2,17 @@ package fuzs.hangglider.init;
 
 import fuzs.hangglider.HangGlider;
 import fuzs.hangglider.capability.GlidingCapability;
-import fuzs.hangglider.world.item.GliderItem;
-import fuzs.hangglider.world.item.GliderType;
+import fuzs.hangglider.world.item.RepairableItem;
 import fuzs.puzzleslib.api.capability.v3.CapabilityController;
 import fuzs.puzzleslib.api.capability.v3.data.EntityCapabilityKey;
 import fuzs.puzzleslib.api.capability.v3.data.SyncStrategy;
 import fuzs.puzzleslib.api.init.v3.registry.RegistryManager;
+import fuzs.puzzleslib.api.init.v3.tags.BoundTagFactory;
 import net.minecraft.core.Holder;
+import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
+import net.minecraft.world.item.Items;
 import net.minecraft.world.item.Rarity;
 
 public class ModRegistry {
@@ -22,14 +24,18 @@ public class ModRegistry {
             () -> new Item(new Item.Properties())
     );
     public static final Holder.Reference<Item> HANG_GLIDER_ITEM = REGISTRY.registerItem("hang_glider",
-            () -> new GliderItem(new Item.Properties().durability(818), GliderType.BASIC)
+            () -> new RepairableItem(() -> Items.LEATHER, new Item.Properties().durability(818))
     );
     public static final Holder.Reference<Item> REINFORCED_HANG_GLIDER_ITEM = REGISTRY.registerItem(
             "reinforced_hang_glider",
-            () -> new GliderItem(new Item.Properties().durability(2202).rarity(Rarity.UNCOMMON),
-                    GliderType.REINFORCED
+            () -> new RepairableItem(() -> Items.PHANTOM_MEMBRANE,
+                    new Item.Properties().durability(2202).rarity(Rarity.UNCOMMON)
             )
     );
+
+    static final BoundTagFactory TAGS = BoundTagFactory.make(HangGlider.MOD_ID);
+    public static final TagKey<Item> HANG_GLIDERS_ITEM_TAG = TAGS.registerItemTag("hang_gliders");
+    public static final TagKey<Item> REINFORCED_HANG_GLIDERS_ITEM_TAG = TAGS.registerItemTag("reinforced_hang_gliders");
 
     static final CapabilityController CAPABILITIES = CapabilityController.from(HangGlider.MOD_ID);
     public static final EntityCapabilityKey<Player, GlidingCapability> GLIDING_CAPABILITY = CAPABILITIES.registerEntityCapability(